requirements.txt
flask>=3.0,<4.0
flask-wtf>=1.2,<2.0
firebase-admin>=6.0,<7.0
python-dotenv>=1.0,<2.0
